Split protocol negotiation into focused modules
